Potentially fixed the executable issue
Former-commit-id: 91a83a65e54012642668e55989f1cdbdf71ebebf

+++ scripts/extract_xtb.py
@@ -3,13 +3,25 @@ import subprocess
 import shlex
 
 if not os.path.isdir("/binaries/xtb"):
-    subprocess.call(shlex.split("tar xvfJ xtb_suite.tar.xz"), cwd="/binaries/")
-    subprocess.call(
+    ret = subprocess.call(shlex.split("tar xvf xtb_suite.tar.xz"), cwd="/binaries/")
+    if ret != 0:
+        raise Exception(f"Got ret {ret}")
+
+    ret = subprocess.call(
         shlex.split("chmod +x crest stda xtb4stda xtb/bin/xtb"),
         cwd="/binaries/",
     )
+    if ret != 0:
+        raise Exception(f"Got ret {ret}")
+
     os.remove("/binaries/xtb_suite.tar.xz")
 if not os.path.isdir("/binaries/Multiwfn"):
-    subprocess.call(shlex.split("tar xvfj Multiwfn.tar.bz2"), cwd="/binaries/")
-    subprocess.call(shlex.split("chmod +x Multiwfn"), cwd="/binaries/")
+    ret = subprocess.call(shlex.split("tar xvfj Multiwfn.tar.bz2"), cwd="/binaries/")
+    if ret != 0:
+        raise Exception(f"Got ret {ret}")
+
+    ret = subprocess.call(shlex.split("chmod +x Multiwfn"), cwd="/binaries/")
+    if ret != 0:
+        raise Exception(f"Got ret {ret}")
+
     os.remove("/binaries/Multiwfn.tar.bz2")
